--- source/containers/App.js
@@ -1,3 +1,0 @@
-if (process.env.NODE_ENV === 'production')
-	module.exports = require('./App.prod')
-else module.exports = require('./App.dev')
