--- source/containers/App.prod.js
@@ -1,18 +1,0 @@
-// App.dev without the redux DevTools
-import React, { Component } from 'react'
-
-import { Provider } from 'react-redux'
-import Layout from './Layout'
-
-export default class App extends Component {
-	render() {
-		const { store } = this.props
-		return (
-			<Provider store={store}>
-				<div>
-					<Layout />
-				</div>
-			</Provider>
-		)
-	}
-}
