+++ source/containers/Layout.css
@@ -10,7 +10,7 @@ body {
 	background: white;
 }
 #js,
-#js > div {
+#dev {
 	height: 100%;
 }
